+++ Example/XcodeEdit-Example.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Example.xcodeproj">
+      location = "self:/Users/tom/Projects/XcodeEdit/Example/XcodeEdit-Example.xcodeproj">
    </FileRef>
 </Workspace>
